Refactor function files
