+++ src/app/functions/capitalize.functions.ts
